+++ rustdoc-json/CHANGELOG.md
@@ -1,6 +1,6 @@
-## Unreleased
-* Add `Builder::package_target()` and `PackageTarget`
+## v0.7.4
 * Correctly determine json path for `Builder::default().package("crate@1.0.0")`
+* Add `Builder::package_target()` and `PackageTarget`
 * Add `Builder::silent()` to suppress stdout and stderr
 
 ## v0.7.3